fix codestyle, use is not instead of ambiguous !=
